refactor: dato post query 리팩토링
