+++ package.json
@@ -10,11 +10,16 @@
     "clean": "rm -rf .next"
   },
   "dependencies": {
+    "@apollo/react-hooks": "^4.0.0",
+    "apollo-boost": "^0.4.9",
+    "apollo-client": "^2.6.10",
     "axios": "^1.5.1",
     "encoding": "^0.1.13",
+    "graphql": "^16.8.1",
     "html-react-parser": "^4.2.2",
     "junyeol-components": "^0.0.51",
     "next": "latest",
+    "next-with-apollo": "^5.3.0",
     "react": "latest",
     "react-dom": "latest",
     "react-markdown": "^9.0.0",
